# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for EasyFind application.",
        "title": "EasyFind API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/claims": {
            "post": {
                "description": "用户申请认领物品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Claim (User)"
                ],
                "summary": "申请认领",
                "parameters": [
                    {
                        "description": "申请信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CreatClaimRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "申请成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/claims/{id}": {
            "get": {
                "description": "根据ID获取认领详情 (发布者和申请者可见)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Claim (User)"
                ],
                "summary": "获取认领详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "认领ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/easyfind_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/easyfind_internal_models.Claim"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/claims/{id}/confirm": {
            "put": {
                "description": "确认认领，将物品状态标记为已找到 (仅发布者可用)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Claim (User)"
                ],
                "summary": "确认认领",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "认领ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "确认成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/login": {
            "post": {
                "description": "用户登录接口，支持学生/老师、失物招领管理员、系统管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/my/claims": {
            "get": {
                "description": "获取当前用户的所有认领申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Claim (User)"
                ],
                "summary": "获取我的认领列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/easyfind_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/register": {
            "post": {
                "description": "注册新账号，用于测试不同身份",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "注册账号 (测试用)",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/password": {
            "put": {
                "description": "修改当前登录用户的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改失败",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "easyfind_internal_models.Account": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "first_login": {
                    "description": "是否首次登录",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "1:学生/老师, 2:失物招领管理员, 3:系统管理员",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyfind_internal_models.UserRole"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "学号或工号",
                    "type": "string"
                }
            }
        },
        "easyfind_internal_models.Claim": {
            "type": "object",
            "properties": {
                "claimant": {
                    "$ref": "#/definitions/easyfind_internal_models.Account"
                },
                "claimant_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "item": {
                    "$ref": "#/definitions/easyfind_internal_models.Item"
                },
                "item_id": {
                    "type": "integer"
                },
                "proof": {
                    "description": "证明材料 (如图片URL)",
                    "type": "string"
                },
                "status": {
                    "description": "状态: pending/approved/rejected/archived",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "easyfind_internal_models.Item": {
            "type": "object",
            "properties": {
                "campus": {
                    "description": "类型 (lost: 失物, found: 招领)",
                    "type": "string"
                },
                "category": {
                    "description": "分类",
                    "type": "string"
                },
                "contact_name": {
                    "description": "联系人姓名",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "联系人电话",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img1": {
                    "description": "照片 (JSON或逗号分隔的URL)",
                    "type": "string"
                },
                "img2": {
                    "description": "照片 (JSON或逗号分隔的URL)",
                    "type": "string"
                },
                "img3": {
                    "description": "照片 (JSON或逗号分隔的URL)",
                    "type": "string"
                },
                "img4": {
                    "type": "string"
                },
                "is_bounty": {
                    "description": "是否有悬赏(保留字段)",
                    "type": "boolean"
                },
                "location": {
                    "description": "地点",
                    "type": "string"
                },
                "process_method": {
                    "description": "处理方式",
                    "type": "string"
                },
                "publisher": {
                    "description": "发布者信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyfind_internal_models.Account"
                        }
                    ]
                },
                "publisher_id": {
                    "type": "integer"
                },
                "reject_reason": {
                    "description": "审核拒绝原因",
                    "type": "string"
                },
                "reviewer": {
                    "description": "审核者信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyfind_internal_models.Account"
                        }
                    ]
                },
                "reviewer_id": {
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyfind_internal_models.ItemStatus"
                        }
                    ]
                },
                "time": {
                    "description": "时间",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/easyfind_internal_models.ItemType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "easyfind_internal_models.ItemStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "matched",
                "claimed",
                "rejected",
                "cancelled",
                "archived"
            ],
            "x-enum-comments": {
                "StatusApproved": "已通过",
                "StatusArchived": "已归档（认领成功或者超时未认领）",
                "StatusCancelled": "已取消",
                "StatusClaimed": "已认领（确认是对的人）",
                "StatusMatched": "已匹配（找到对应的人来对话了）",
                "StatusPending": "待审核",
                "StatusRejected": "已驳回"
            },
            "x-enum-descriptions": [
                "待审核",
                "已通过",
                "已匹配（找到对应的人来对话了）",
                "已认领（确认是对的人）",
                "已驳回",
                "已取消",
                "已归档（认领成功或者超时未认领）"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusApproved",
                "StatusMatched",
                "StatusClaimed",
                "StatusRejected",
                "StatusCancelled",
                "StatusArchived"
            ]
        },
        "easyfind_internal_models.ItemType": {
            "type": "string",
            "enum": [
                "lost",
                "found"
            ],
            "x-enum-varnames": [
                "TypeLost",
                "TypeFound"
            ]
        },
        "easyfind_internal_models.UserRole": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "RoleStudentTeacher",
                "RoleLFAdmin",
                "RoleSysAdmin"
            ]
        },
        "easyfind_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "使用 data 统一包裹数据"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_controllers.CreatClaimRequest": {
            "type": "object",
            "required": [
                "item_id"
            ],
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "proof": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "role",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "remember": {
                    "description": "记住我",
                    "type": "boolean"
                },
                "role": {
                    "description": "1:学生/老师, 2:失物招领管理员, 3:系统管理员",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "role",
                "username"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}